+++ QCM_main/UISettings.py
@@ -157,7 +157,7 @@ settings_init = {
         'label_settings_mechanics_refG',
         'comboBox_settings_mechanics_refG',
         # 'comboBox_settings_mechanics_selectmodel',
-        'groupBox_settings_mechanics_mech_film',
+        # 'groupBox_settings_mechanics_mech_film',
 
         # 'groupBox_settings_mechanics_contour',
         # 'pushButton_settings_mechanics_simulator',
@@ -272,16 +272,17 @@ settings_init = {
         'delg_calcs':        r'$\Delta\Gamma$ (Hz)',
         'delg_exps':         r'$\Delta\Gamma_{exp}$ (Hz)',
         'drho':              r'd$\rho$ ($\mu$m$\cdot$g/cm$^3$)',
-        'grhos':           r'$|G_{n}^*|\rho$ (Pa$\cdot$g/cm$^3$)',
+        'grhos':             r'$|G_{n}^*|\rho$ (Pa$\cdot$g/cm$^3$)',
         'phi':               r'$\phi$ ($\degree$)',
-        'dlams':           r'd/$\lambda_{n}$',
-        'lamrhos':            r'$\lambda\rho$ ($\mu$m$\cdot$g/cm$^3$)',
-        'delrhos':            r'$\delta\rho$ ($\mu$m$\cdot$g/cm$^3$)',
-        'delf_delfsns':       r'$\Delta$f/$\Delta$f$_{sn}$',
+        'etarhos':           r'$|\eta_{n}^*|\rho$ (Pa$\cdot$s$\cdot$g/cm$^3$)',
+        'dlams':             r'd/$\lambda_{n}$',
+        'lamrhos':           r'$\lambda\rho$ ($\mu$m$\cdot$g/cm$^3$)',
+        'delrhos':           r'$\delta\rho$ ($\mu$m$\cdot$g/cm$^3$)',
+        'delf_delfsns':      r'$\Delta$f/$\Delta$f$_{sn}$',
         'delg_delfsn_exps':  r'$(\Delta\Gamma$/$\Delta$f$_{sn})_{exp}$',
         'delg_delfsn_calcs': r'$\Delta\Gamma$/$\Delta$f$_{sn}$',
-        'rh_exp':           r'r$_{h,exp}$',
-        'rh_calc':          r'r$_h$',
+        'rh_exp':            r'r$_{h,exp}$',
+        'rh_calc':           r'r$_h$',
         'rd_exps':           r'r$_{d,exp}$',
         'rd_calcs':          r'r$_d$',
     },
@@ -294,14 +295,15 @@ settings_init = {
         'delg_exps':         u'\u0394\u0393' + ' (Hz)', # ΔΓ (Hz)
         'delg_calcs':        u'\u0394\u0393' + 'calc (Hz)', # ΔΓcalc (Hz)
         'drho':              'd' + u'\u03C1' + ' (' + u'\u03BC' + 'm' + u'\u2219' 'g/cm'+ u'\u00B3' + ')', # dρ (μm∙g/m³)
-        'grhos':           '|G*|' + u'\u03C1' + ' (Pa' + u'\u2219' + 'g/cm' + u'\u00B3' + ')', # |G*|ρ (Pa∙g/cm³)
-        'phi':                u'\u03A6' + ' (' + u'\u00B0' + ')', # Φ (°)
-        'dlams':           'd/' + u'\u03BB\u2099', # d/λₙ
-        'lamrhos':            u'\u03BB\u03C1' + ' (' + u'\u03BC' + 'm' + u'\u2219' 'g/cm'+ u'\u00B3' + ')', # λρ (μm∙g/m³)
-        'delrhos':            u'\u03B4\u03C1' + ' (' + u'\u03BC' + 'm' + u'\u2219' 'g/cm'+ u'\u00B3' + ')', # δρ (μm∙g/m³)
-        'delf_delfsns':       u'\u0394' + 'f/' + u'\u0394' + 'f' + u'\u209B\u2099', # Δf/Δfₛₙ
+        'grhos':             '|G*|' + u'\u03C1' + ' (Pa' + u'\u2219' + 'g/cm' + u'\u00B3' + ')', # |G*|ρ (Pa∙g/cm³)
+        'phi':               u'\u03A6' + ' (' + u'\u00B0' + ')', # Φ (°)
+        'etarhos':           '|' + u'\u03B7' + '*|' + u'\u03C1' + ' (Pa' + u'\u2219' + 's' + u'\u2219' + 'g/cm' + u'\u00B3' + ')', # |η*|ρ (Pa∙s∙g/cm³)
+        'dlams':             'd/' + u'\u03BB\u2099', # d/λₙ
+        'lamrhos':           u'\u03BB\u03C1' + ' (' + u'\u03BC' + 'm' + u'\u2219' 'g/cm'+ u'\u00B3' + ')', # λρ (μm∙g/m³)
+        'delrhos':           u'\u03B4\u03C1' + ' (' + u'\u03BC' + 'm' + u'\u2219' 'g/cm'+ u'\u00B3' + ')', # δρ (μm∙g/m³)
+        'delf_delfsns':      u'\u0394' + 'f/' + u'\u0394' + 'f' + u'\u209B\u2099', # Δf/Δfₛₙ
         'delg_delfsn_calcs': u'\u0394\u0393' + '/' + u'\u0394' + 'f' + u'\u209B\u2099', # ΔΓ/Δfₛₙ
-        'rh_calc':                'rh',
+        'rh_calc':           'rh',
         'rd_calcs':          'rd',
         't':                 'Time (s)', # Time (s)
         'temp':              'Temp. (' + u'\u00B0' + 'C)', # Temp. (°C)
@@ -336,6 +338,8 @@ settings_init = {
         ('1', 'ADC 1'),
         ('2', 'ADC 2'),
     ]),
+
+    'vna_wait_time_extra': 0.05, # extra time adds to wait_time
 
     'ref_channel_opts': OrderedDict([
     # key: str(number); val: for display in combobox
@@ -437,22 +441,22 @@ settings_init = {
 
     # minimum/maximum layer for mechanic property calculation
     'min_mech_layers': 0,
-    'max_mech_layers': 1,
+    'max_mech_layers': 2,
 
     # options for comboBox_settings_mechanics_selectmodel
     'qcm_model_opts': {
-        'onelayer': 'One layer',
-        'bulk': 'Bulk',
-        'twolayers': 'Two layers',
+        'onelayer': 'In air',
+        # 'bulk': 'Bulk',
+        'twolayers': 'In medium',
     },
 
     # calctype
     'calctype_opts':{
-        'SLA': 'SLA',
-        'LL': 'LL',
+        'SLA': 'Small-load approximation',
+        'LL': 'Large-load',
     },
 
-    'qcm_layer_known_source_opts': {
+    'qcm_layer_source_opts': {
         'ind': 'Index',
         'prop': 'Prop.',
         # 'fg': u'\u0394' + 'f&' + u'\u0394\u0393',
@@ -461,26 +465,29 @@ settings_init = {
         # 'none': '--',
     },
 
-    'qcm_layer_unknown_source_opts': {
-        'none': '--',
-        'ind': 'Index',
-        'guess': 'Guess',
-    },
+    'mechanics_modeswitch': 0, 
 
-    'qcm_layer_bulk_name_opts': {
-        'air': {
-            'drho': np.inf,
-            'grho': 0,
-            'phi': 0,
-            'rh': 3,
-        },
-        'water': {
-            'drho': np.inf, 
-            'grho': 1e5, # in Pa
-            'phi': np.pi /2,
-            'rh': 3,
-        },
-    },
+
+    # 'qcm_layer_unknown_source_opts': {
+    #     'none': '--',
+    #     'ind': 'Index',
+    #     'guess': 'Guess',
+    # },
+
+    # 'qcm_layer_bulk_name_opts': {
+    #     'air': {
+    #         'drho': np.inf,
+    #         'grho': 0,
+    #         'phi': 0,
+    #         'rh': 3,
+    #     },
+    #     'water': {
+    #         'drho': np.inf, 
+    #         'grho': 1e5, # in Pa
+    #         'phi': np.pi /2,
+    #         'rh': 3,
+    #     },
+    # },
 
     # steps ofr span control slider
     'span_ctrl_steps': [1, 2, 5, 10, 20, 50, 100],
@@ -624,6 +631,9 @@ settings_default = {
     # 'vna_path': r'C:/...../myVNA.exe'.
     # if this key is empty, the program will look for the file in the default list in settings_init['vna_path']
     'vna_path': r'',
+    # keep key below (vna_wait_time_extra) commented.
+    # it can be actived in user setting file
+    # 'vna_wait_time_extra': 0.05, # in s. This extra time will be added to the calculated value
 
     # default checkbox harm states (checkbox with False value can be omit from this list)
     'checkBox_harm1': True,
@@ -636,9 +646,9 @@ settings_default = {
     # NOTE: keep this key commented
     # 'dateTimeEdit_reftime': '2000-01-01 00:00:00.000',
     
-    'lineEdit_recordinterval': 5,
-    'lineEdit_refreshresolution': 1,
-    'lineEdit_scaninterval': 5,
+    'spinBox_recordinterval': 15,
+    'spinBox_refreshresolution': 1,
+    'spinBox_scaninterval': 15,
 
     # default fitting and display options
     'checkBox_dynamicfit': True,
@@ -703,8 +713,8 @@ settings_default = {
 
     ### default plot selections ###
     # default selections for spectra show
-    'radioButton_spectra_showGp': True,
-    'radioButton_spectra_showBp': False,
+    'radioButton_spectra_showGp': False,
+    'radioButton_spectra_showBp': True,
     'radioButton_spectra_showpolar': False,
     'checkBox_spectra_showchi': False,
 
@@ -752,7 +762,7 @@ settings_default = {
     'spinBox_settings_mechanics_nhcalc_n3': 3,
 
     'comboBox_settings_mechanics_refG': '3', # reference harmonic for property
-    'spinBox_mech_expertmode_layernum': 1, # number of layers for expert mode mechanic 
+    'spinBox_mech_expertmode_layernum': 2, # number of layers for expert mode mechanic 
 
     'comboBox_settings_mechanics_calctype': 'LL', # 'LL' or 'SLA'
 
@@ -771,14 +781,14 @@ harm_tree = {
     'comboBox_tracking_condition': 'auto',
     # default fit settings
     'checkBox_harmfit': True,
-    'spinBox_harmfitfactor': 3,
+    'spinBox_harmfitfactor': 1,
     'spinBox_peaks_num': 1, 
     'radioButton_peaks_num_max': True,
     'radioButton_peaks_num_fixed': False,
     'radioButton_peaks_policy_minf': False,
     'radioButton_peaks_policy_maxamp': True,
     'checkBox_settings_settings_harmzerophase': False,
-    'lineEdit_peaks_threshold': 0.0001,
+    'lineEdit_peaks_threshold': 0.00001,
     'lineEdit_peaks_prominence': 0.0001,
 }