+++ QCM_main/modules/TempDevices.py
@@ -47,7 +47,6 @@ def device_info(devtype):
     '''
     return a single device 
     '''
-    print('devtype', devtype) #testprint
     devices = list_devices()
     for device in devices:
     #     if device.product_type != devtype: